+++ packages/typespec-rust/src/codemodel/client.ts
@@ -234,7 +234,7 @@ export interface PartialBodyParameter extends HTTPParameterBase {
   kind: 'partialBody';
 
   /** the model in which the partial param is placed */
-  type: types.RequestContent<types.Model>;
+  type: types.RequestContent<types.Payload<types.Model>>;
 
   /** the name of the field over the wire in model.fields for this param */
   serde: string;
@@ -494,7 +494,7 @@ export class PageableStrategyNextLink implements PageableStrategyNextLink {
 }
 
 export class PartialBodyParameter extends HTTPParameterBase implements PartialBodyParameter {
-  constructor(name: string, location: ParameterLocation, optional: boolean, serde: string, type: types.RequestContent<types.Model>) {
+  constructor(name: string, location: ParameterLocation, optional: boolean, serde: string, type: types.RequestContent<types.Payload<types.Model>>) {
     super(name, location, optional, type);
     this.kind = 'partialBody';
     this.serde = serde;